[Feat] 네트워크 레이어 작성
